+++ src/collective/nitf/config.py
@@ -193,4 +193,3 @@ URGENCIES = SimpleVocabulary([
 DEFAULT_GENRE = u'Current'
 DEFAULT_SECTION = _(u'General')
 DEFAULT_URGENCY = NORMAL
-DEFAULT_RELATABLE_CONTENT_TYPES = [u'collective.nitf.content']